# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: AgentOps Desktop HTTP API
  description: |
    HTTP REST API for AgentOps Desktop — manage AI agents, goals, tasks, schedules, squads, cost budgets, adapters, and skills.

    **Base URL:** `http://localhost:3967`
    **Authentication:** Bearer token via `POST /auth/login`. Include as `Authorization: Bearer <token>` on all `/api/*` routes.
    **Default port:** 3967 (configurable via `PORT` env var)
  version: "0.2.0"
  contact:
    name: AgentOps Team
  license:
    name: UNLICENSED
servers:
  - url: "http://localhost:3967"
    description: Local HTTP API server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
            status:
              type: integer
              example: 404
    Agent:
      type: object
      properties:
        id:
          type: string
          example: "agent-1"
        name:
          type: string
          example: "claude-dev"
        type:
          type: string
          enum: [autonomous, supervised, manual]
          example: autonomous
        status:
          type: string
          enum: [idle, running, paused, error]
          example: idle
        command:
          type: string
          nullable: true
        execPath:
          type: string
          nullable: true
          example: /usr/local/bin/claude
        cwd:
          type: string
          nullable: true
          example: /project
        createdAt:
          type: integer
          example: 1716900000000
        updatedAt:
          type: integer
          example: 1716900000000
    Goal:
      type: object
      properties:
        id:
          type: string
          example: "goal-1"
        title:
          type: string
          example: Ship MVP
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, archived]
          example: active
        taskIds:
          type: array
          items:
            type: string
        createdAt:
          type: integer
        updatedAt:
          type: integer
    Task:
      type: object
      properties:
        id:
          type: string
          example: "task-1"
        title:
          type: string
          example: Implement auth
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, assigned, running, done, failed, blocked]
          example: pending
        goalId:
          type: string
          nullable: true
        assigneeAgentId:
          type: string
          nullable: true
        dependsOn:
          type: object
          nullable: true
        output:
          type: object
          nullable: true
        createdAt:
          type: integer
        updatedAt:
          type: integer
    LogEntry:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
          nullable: true
        taskId:
          type: string
          nullable: true
        message:
          type: string
        level:
          type: string
          enum: [debug, info, warn, error]
        stream:
          type: string
          enum: [stdout, stderr]
        timestamp:
          type: integer
    Schedule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        cronExpr:
          type: string
          example: "0 9 * * *"
        goalId:
          type: string
          nullable: true
        agentId:
          type: string
          nullable: true
        taskTemplate:
          type: object
          nullable: true
        maxExecutions:
          type: integer
          nullable: true
        enabled:
          type: boolean
        nextRunAt:
          type: string
          format: date-time
        createdAt:
          type: integer
        updatedAt:
          type: integer
    Squad:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        leaderId:
          type: string
          nullable: true
        status:
          type: string
          enum: [idle, running, error]
        members:
          type: array
          items:
            type: object
            properties:
              agentId:
                type: string
              role:
                type: string
                enum: [member, leader]
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    Adapter:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          example: claude
        name:
          type: string
        classPath:
          type: string
          nullable: true
        config:
          type: object
          nullable: true
        enabled:
          type: boolean
        createdAt:
          type: integer
        updatedAt:
          type: integer
    Skill:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        content:
          type: string
        tags:
          type: object
          nullable: true
        createdAt:
          type: integer
        updatedAt:
          type: integer
paths:
...
  /api/agents:
    get:
      summary: List agents
      operationId: agentsList
      tags: [Agents]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [idle, running, paused, error] }
        - name: sortBy
          in: query
          schema: { type: string, enum: [name, status, createdAt, updatedAt] }
        - name: sortOrder
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
    post:
      summary: Create agent
      operationId: agentsCreate
      tags: [Agents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                type: { type: string, enum: [autonomous, supervised, manual] }
                command: { type: string, maxLength: 1000 }
                execPath: { type: string, maxLength: 1000 }
                cwd: { type: string, maxLength: 500 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Agent" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/goals:
    get:
      summary: List goals
      operationId: goalsList
      tags: [Goals]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [active, completed, archived] } }
        - { name: sortBy, in: query, schema: { type: string, enum: [title, status, createdAt, updatedAt] } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        "200":
          description: Goal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Goal" }
    post:
      summary: Create goal
      operationId: goalsCreate
      tags: [Goals]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 500 }
                description: { type: string, maxLength: 5000 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Goal" }
...
  /api/tasks:
    get:
      summary: List tasks
      operationId: tasksList
      tags: [Tasks]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [pending, assigned, running, done, failed, blocked] } }
        - { name: goalId, in: query, schema: { type: string } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, updatedAt, title, status] } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Task" }
    post:
      summary: Create task
      operationId: tasksCreate
      tags: [Tasks]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 500 }
                description: { type: string, maxLength: 5000 }
                goalId: { type: string }
                assigneeAgentId: { type: string }
                dependsOn: { type: object }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Task" }
...
  /api/logs:
    get:
      summary: List log entries
      operationId: logsList
      tags: [Logs]
      security:
        - bearerAuth: []
      parameters:
        - { name: agentId, in: query, schema: { type: string } }
        - { name: taskId, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/LogEntry" }
    post:
      summary: Append log entry
      operationId: logsAppend
      tags: [Logs]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                agentId: { type: string }
                taskId: { type: string }
                message: { type: string }
                level: { type: string, enum: [debug, info, warn, error] }
                stream: { type: string, enum: [stdout, stderr] }
      responses:
        "201":
          description: Created

  # ── Stats ──
...
  /api/settings:
    get:
      summary: Get all settings
      operationId: settingsGet
      tags: [Settings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
    patch:
      summary: Update settings
      operationId: settingsUpdate
      tags: [Settings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings: { type: object }
      responses:
        "200":
          description: Updated settings

  # ── Schedules ──
  /api/schedules:
    get:
      summary: List schedules
      operationId: schedulesList
      tags: [Schedules]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: Schedule list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Schedule" }
                  total: { type: integer }
    post:
      summary: Create schedule
      operationId: schedulesCreate
      tags: [Schedules]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, cronExpr]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                cronExpr: { type: string, minLength: 1 }
                goalId: { type: string }
                agentId: { type: string }
                taskTemplate: { type: object }
                maxExecutions: { type: integer }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Schedule" }
...
  /api/squads:
    get:
      summary: List squads
      operationId: squadsList
      tags: [Squads]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [idle, running, error] } }
      responses:
        "200":
          description: Squad list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Squad" }
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      offset: { type: integer }
                      limit: { type: integer }
    post:
      summary: Create squad
      operationId: squadsCreate
      tags: [Squads]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                description: { type: string, maxLength: 1000 }
                leaderId: { type: string }
                members:
                  type: array
                  items:
                    oneOf:
                      - type: string
                      - type: object
                        properties:
                          agentId: { type: string }
                          role: { type: string, enum: [member, leader] }
      responses:
        "201":
          description: Created
        "404":
          description: Not found
...
  /api/adapters:
    get:
      summary: List adapters
      operationId: adaptersList
      tags: [Adapters]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: enabled, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Adapter list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Adapter" }
                  total: { type: integer }
    post:
      summary: Create adapter
      operationId: adaptersCreate
      tags: [Adapters]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, minLength: 1, maxLength: 100 }
                name: { type: string, maxLength: 200 }
                classPath: { type: string }
                config: { type: object }
                enabled: { type: boolean }
      responses:
        "201":
          description: Created
        "409":
          description: Adapter type already exists
...
  /api/skills:
    get:
      summary: List skills
      operationId: skillsList
      tags: [Skills]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: tags, in: query, schema: { type: string, description: "Comma-separated tags" } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Skill list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Skill" }
    post:
      summary: Create skill
      operationId: skillsCreate
      tags: [Skills]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, content]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                description: { type: string, maxLength: 2000 }
                content: { type: string, minLength: 1 }
                tags: { type: object }
      responses:
        "201":
          description: Created
        "409":
          description: Skill name already exists
...